Continue updating about page
